feat(api): bind http server with rpc server
