+++ api/api.go
@@ -42,7 +42,6 @@
 // @scope.admin Grants read and write access to administrative information
 
 // @x-extension-openapi {"example": "value on a json format"}
-
 package api
 
 import (
@@ -50,10 +49,11 @@ import (
 	"github.com/swaggo/files"
 	"github.com/swaggo/gin-swagger"
 
-	"./docs" // docs is generated by Swag CLI, you have to import it.
+	"github.com/darwinia-network/shadow/api/docs"
+	"github.com/darwinia-network/shadow/internal/util"
 )
 
-func main() {
+func Swagger() {
 	// programatically set swagger info
 	docs.SwaggerInfo.Title = "Swagger Example API"
 	docs.SwaggerInfo.Description = "This is a sample server Petstore server."
@@ -63,22 +63,21 @@ func main() {
 	docs.SwaggerInfo.Schemes = []string{"http", "https"}
 
 	r := gin.Default()
-	c := controller.NewController()
+	// c := NewController()
 
-	v1 := r.Group("/api/v1")
-	{
-		accounts := v1.Group("/accounts")
-		{
-			accounts.GET(":id", c.ShowAccount)
-			accounts.GET("", c.ListAccounts)
-			accounts.POST("", c.AddAccount)
-			accounts.DELETE(":id", c.DeleteAccount)
-			accounts.PATCH(":id", c.UpdateAccount)
-			accounts.POST(":id/images", c.UploadAccountImage)
-		}
-    //...
-	}
+	// v1 := r.Group("/api/v1")
+	// {
+	// 	accounts := v1.Group("/accounts")
+	// 	{
+	// 		accounts.GET(":id", c.ShowAccount)
+	// 		accounts.GET("", c.ListAccounts)
+	// 		accounts.POST("", c.AddAccount)
+	// 		accounts.DELETE(":id", c.DeleteAccount)
+	// 		accounts.PATCH(":id", c.UpdateAccount)
+	// 		accounts.POST(":id/images", c.UploadAccountImage)
+	// 	}
+	//
+	// }
 	r.GET("/swagger/*any", ginSwagger.WrapHandler(swaggerFiles.Handler))
-	r.Run(":8080")
+	util.Assert(r.Run(":8080"))
 }
-//...
